+++ src/SqlArtisan/Internal/SqlPart/Clause/Update/UpdateSetClause.cs
@@ -19,7 +19,12 @@ internal sealed class UpdateSetClause : SqlPart
 
         for (int i = 0; i < items.Length; i++)
         {
-            if (items[i] is not EqualityCondition)
+            if (items[i] is null)
+            {
+                throw new ArgumentNullException(
+                    nameof(items), ExpressionResolver.NullValueMessage);
+            }
+            else if (items[i] is not EqualityCondition)
             {
                 throw new ArgumentException(
                     $"Invalid type for EqualityCondition: {items[i].GetType()}");